# opt: faster skip in ast