--- a/.github/workflows/benchmark-linux-arm64.yml
+++ b/.github/workflows/benchmark-linux-arm64.yml
@@ -6,6 +6,9 @@ jobs:
   build:
     runs-on: [arm]
     steps:
+      - name: Clear repository
+        run: sudo rm -fr $GITHUB_WORKSPACE && mkdir $GITHUB_WORKSPACE
+
       - uses: actions/checkout@v2
 
       - name: Check Branch