--- a/ast/search.go
+++ b/ast/search.go
@@ -16,12 +16,6 @@
 
 package ast
 
-import (
-    `fmt`
-
-    `github.com/bytedance/sonic/internal/native/types`
-)
-
 type Searcher struct {
     parser Parser
 }
@@ -33,40 +27,4 @@ func NewSearcher(str string) *Searcher {
             noLazy: false,
         },
     }
-}
-
-func (self *Searcher) GetByPath(path ...interface{}) (Node, error) {
-    self.parser.p = 0
-
-    var err types.ParsingError
-    for _, p := range path {
-        switch p := p.(type) {
-        case int:
-            if err = self.parser.searchIndex(p); err != 0 {
-                return Node{}, self.parser.ExportError(err)
-            }
-        case string:
-            if err = self.parser.searchKey(p); err != 0 {
-                return Node{}, self.parser.ExportError(err)
-            }
-        default:
-            panic("path must be either int or string")
-        }
-    }
-
-    var start = self.parser.p
-    if start, err = self.parser.skip(); err != 0 {
-        return Node{}, self.parser.ExportError(err)
-    }
-    ns := len(self.parser.s)
-    if self.parser.p > ns || start >= ns || start>=self.parser.p {
-        return Node{}, fmt.Errorf("skip %d char out of json boundary", start)
-    }
-
-    t := switchRawType(self.parser.s[start])
-    if t == _V_NONE {
-        return Node{}, self.parser.ExportError(err)
-    }
-
-    return newRawNode(self.parser.s[start:self.parser.p], t), nil
 }