--- a/external_jsonlib_test/go.mod
+++ b/external_jsonlib_test/go.mod
@@ -3,6 +3,7 @@ module github.com/bytedance/sonic/external_jsonlib_test
 go 1.18
 
 require (
+	github.com/buger/jsonparser v1.1.1
 	github.com/bytedance/sonic v1.4.0
 	github.com/goccy/go-json v0.9.11
 	github.com/json-iterator/go v1.1.12