--- a/internal/native/avx/native_export_amd64.go
+++ b/internal/native/avx/native_export_amd64.go
@@ -41,7 +41,9 @@ var (
 
 var (
     S_skip_one    = _subr__skip_one
+    S_skip_one_fast = _subr__skip_one_fast
     S_skip_array  = _subr__skip_array
     S_skip_object = _subr__skip_object
     S_skip_number = _subr__skip_number
+    S_get_by_path = _subr__get_by_path
 )