--- a/internal/native/avx/native_subr_amd64.go
+++ b/internal/native/avx/native_subr_amd64.go
@@ -9,29 +9,32 @@ package avx
 func __native_entry__() uintptr
 
 var (
-    _subr__f32toa       = __native_entry__() + 24592
-    _subr__f64toa       = __native_entry__() + 496
-    _subr__html_escape  = __native_entry__() + 10480
-    _subr__i64toa       = __native_entry__() + 4176
-    _subr__lspace       = __native_entry__() + 80
-    _subr__quote        = __native_entry__() + 5552
-    _subr__skip_array   = __native_entry__() + 22864
-    _subr__skip_number  = __native_entry__() + 24336
-    _subr__skip_object  = __native_entry__() + 22912
-    _subr__skip_one     = __native_entry__() + 20992
-    _subr__u64toa       = __native_entry__() + 4288
-    _subr__unquote      = __native_entry__() + 7296
-    _subr__validate_one = __native_entry__() + 24480
-    _subr__value        = __native_entry__() + 13728
-    _subr__vnumber      = __native_entry__() + 18736
-    _subr__vsigned      = __native_entry__() + 20288
-    _subr__vstring      = __native_entry__() + 15808
-    _subr__vunsigned    = __native_entry__() + 20640
+    _subr__f32toa        = __native_entry__() + 28656
+    _subr__f64toa        = __native_entry__() + 496
+    _subr__get_by_path   = __native_entry__() + 26848
+    _subr__html_escape   = __native_entry__() + 10480
+    _subr__i64toa        = __native_entry__() + 4176
+    _subr__lspace        = __native_entry__() + 80
+    _subr__quote         = __native_entry__() + 5552
+    _subr__skip_array    = __native_entry__() + 20160
+    _subr__skip_number   = __native_entry__() + 23472
+    _subr__skip_object   = __native_entry__() + 22048
+    _subr__skip_one      = __native_entry__() + 23616
+    _subr__skip_one_fast = __native_entry__() + 23824
+    _subr__u64toa        = __native_entry__() + 4288
+    _subr__unquote       = __native_entry__() + 7296
+    _subr__validate_one  = __native_entry__() + 23648
+    _subr__value         = __native_entry__() + 13728
+    _subr__vnumber       = __native_entry__() + 17904
+    _subr__vsigned       = __native_entry__() + 19456
+    _subr__vstring       = __native_entry__() + 15808
+    _subr__vunsigned     = __native_entry__() + 19808
 )
 
 const (
     _stack__f32toa = 64
     _stack__f64toa = 80
+    _stack__get_by_path = 296
     _stack__html_escape = 64
     _stack__i64toa = 16
     _stack__lspace = 8
@@ -40,6 +43,7 @@ const (
     _stack__skip_number = 72
     _stack__skip_object = 128
     _stack__skip_one = 128
+    _stack__skip_one_fast = 208
     _stack__u64toa = 8
     _stack__unquote = 72
     _stack__validate_one = 128
@@ -53,6 +57,7 @@ const (
 var (
     _ = _subr__f32toa
     _ = _subr__f64toa
+    _ = _subr__get_by_path
     _ = _subr__html_escape
     _ = _subr__i64toa
     _ = _subr__lspace
@@ -61,6 +66,7 @@ var (
     _ = _subr__skip_number
     _ = _subr__skip_object
     _ = _subr__skip_one
+    _ = _subr__skip_one_fast
     _ = _subr__u64toa
     _ = _subr__unquote
     _ = _subr__validate_one
@@ -74,6 +80,7 @@ var (
 const (
     _ = _stack__f32toa
     _ = _stack__f64toa
+    _ = _stack__get_by_path
     _ = _stack__html_escape
     _ = _stack__i64toa
     _ = _stack__lspace
@@ -82,6 +89,7 @@ const (
     _ = _stack__skip_number
     _ = _stack__skip_object
     _ = _stack__skip_one
+    _ = _stack__skip_one_fast
     _ = _stack__u64toa
     _ = _stack__unquote
     _ = _stack__validate_one