--- a/internal/native/avx2/native_amd64.go
+++ b/internal/native/avx2/native_amd64.go
@@ -97,6 +97,11 @@ func __skip_one(s *string, p *int, m *types.StateMachine, flags uint64) (ret int
 //go:nosplit
 //go:noescape
 //goland:noinspection GoUnusedParameter
+func __skip_one_fast(s *string, p *int) (ret int)
+
+//go:nosplit
+//go:noescape
+//goland:noinspection GoUnusedParameter
 func __skip_array(s *string, p *int, m *types.StateMachine, flags uint64) (ret int)
 
 //go:nosplit
@@ -113,3 +118,8 @@ func __skip_number(s *string, p *int) (ret int)
 //go:noescape
 //goland:noinspection GoUnusedParameter
 func __validate_one(s *string, p *int, m *types.StateMachine) (ret int)
+
+//go:nosplit
+//go:noescape
+//goland:noinspection GoUnusedParameter
+func __get_by_path(s *string, p *int, path *[]interface{}) (ret int)