--- a/internal/native/avx2/native_subr_amd64.go
+++ b/internal/native/avx2/native_subr_amd64.go
@@ -9,29 +9,32 @@ package avx2
 func __native_entry__() uintptr
 
 var (
-    _subr__f32toa       = __native_entry__() + 28464
-    _subr__f64toa       = __native_entry__() + 752
-    _subr__html_escape  = __native_entry__() + 12320
-    _subr__i64toa       = __native_entry__() + 4432
-    _subr__lspace       = __native_entry__() + 224
-    _subr__quote        = __native_entry__() + 5904
-    _subr__skip_array   = __native_entry__() + 26112
-    _subr__skip_number  = __native_entry__() + 28208
-    _subr__skip_object  = __native_entry__() + 26160
-    _subr__skip_one     = __native_entry__() + 24208
-    _subr__u64toa       = __native_entry__() + 4544
-    _subr__unquote      = __native_entry__() + 8848
-    _subr__validate_one = __native_entry__() + 28352
-    _subr__value        = __native_entry__() + 16896
-    _subr__vnumber      = __native_entry__() + 21952
-    _subr__vsigned      = __native_entry__() + 23504
-    _subr__vstring      = __native_entry__() + 19280
-    _subr__vunsigned    = __native_entry__() + 23856
+    _subr__f32toa        = __native_entry__() + 32816
+    _subr__f64toa        = __native_entry__() + 752
+    _subr__get_by_path   = __native_entry__() + 30896
+    _subr__html_escape   = __native_entry__() + 12320
+    _subr__i64toa        = __native_entry__() + 4432
+    _subr__lspace        = __native_entry__() + 224
+    _subr__quote         = __native_entry__() + 5904
+    _subr__skip_array    = __native_entry__() + 23472
+    _subr__skip_number   = __native_entry__() + 27440
+    _subr__skip_object   = __native_entry__() + 25392
+    _subr__skip_one      = __native_entry__() + 27584
+    _subr__skip_one_fast = __native_entry__() + 27984
+    _subr__u64toa        = __native_entry__() + 4544
+    _subr__unquote       = __native_entry__() + 8848
+    _subr__validate_one  = __native_entry__() + 27616
+    _subr__value         = __native_entry__() + 16896
+    _subr__vnumber       = __native_entry__() + 21216
+    _subr__vsigned       = __native_entry__() + 22768
+    _subr__vstring       = __native_entry__() + 19280
+    _subr__vunsigned     = __native_entry__() + 23120
 )
 
 const (
     _stack__f32toa = 64
     _stack__f64toa = 80
+    _stack__get_by_path = 304
     _stack__html_escape = 72
     _stack__i64toa = 16
     _stack__lspace = 8
@@ -40,19 +43,21 @@ const (
     _stack__skip_number = 80
     _stack__skip_object = 136
     _stack__skip_one = 136
+    _stack__skip_one_fast = 216
     _stack__u64toa = 8
     _stack__unquote = 72
     _stack__validate_one = 136
     _stack__value = 336
     _stack__vnumber = 248
     _stack__vsigned = 16
-    _stack__vstring = 128
+    _stack__vstring = 136
     _stack__vunsigned = 24
 )
 
 var (
     _ = _subr__f32toa
     _ = _subr__f64toa
+    _ = _subr__get_by_path
     _ = _subr__html_escape
     _ = _subr__i64toa
     _ = _subr__lspace
@@ -61,6 +66,7 @@ var (
     _ = _subr__skip_number
     _ = _subr__skip_object
     _ = _subr__skip_one
+    _ = _subr__skip_one_fast
     _ = _subr__u64toa
     _ = _subr__unquote
     _ = _subr__validate_one
@@ -74,6 +80,7 @@ var (
 const (
     _ = _stack__f32toa
     _ = _stack__f64toa
+    _ = _stack__get_by_path
     _ = _stack__html_escape
     _ = _stack__i64toa
     _ = _stack__lspace
@@ -82,6 +89,7 @@ const (
     _ = _stack__skip_number
     _ = _stack__skip_object
     _ = _stack__skip_one
+    _ = _stack__skip_one_fast
     _ = _stack__u64toa
     _ = _stack__unquote
     _ = _stack__validate_one