--- a/internal/native/sse/native_subr_amd64.go
+++ b/internal/native/sse/native_subr_amd64.go
@@ -9,29 +9,32 @@ package sse
 func __native_entry__() uintptr
 
 var (
-    _subr__f32toa       = __native_entry__() + 24640
-    _subr__f64toa       = __native_entry__() + 464
-    _subr__html_escape  = __native_entry__() + 10416
-    _subr__i64toa       = __native_entry__() + 4048
-    _subr__lspace       = __native_entry__() + 80
-    _subr__quote        = __native_entry__() + 5456
-    _subr__skip_array   = __native_entry__() + 22928
-    _subr__skip_number  = __native_entry__() + 24432
-    _subr__skip_object  = __native_entry__() + 22976
-    _subr__skip_one     = __native_entry__() + 21056
-    _subr__u64toa       = __native_entry__() + 4176
-    _subr__unquote      = __native_entry__() + 7232
-    _subr__validate_one = __native_entry__() + 24576
-    _subr__value        = __native_entry__() + 13680
-    _subr__vnumber      = __native_entry__() + 18800
-    _subr__vsigned      = __native_entry__() + 20352
-    _subr__vstring      = __native_entry__() + 15760
-    _subr__vunsigned    = __native_entry__() + 20704
+    _subr__f32toa        = __native_entry__() + 29152
+    _subr__f64toa        = __native_entry__() + 464
+    _subr__get_by_path   = __native_entry__() + 27392
+    _subr__html_escape   = __native_entry__() + 10416
+    _subr__i64toa        = __native_entry__() + 4048
+    _subr__lspace        = __native_entry__() + 80
+    _subr__quote         = __native_entry__() + 5456
+    _subr__skip_array    = __native_entry__() + 20144
+    _subr__skip_number   = __native_entry__() + 23488
+    _subr__skip_object   = __native_entry__() + 22032
+    _subr__skip_one      = __native_entry__() + 23632
+    _subr__skip_one_fast = __native_entry__() + 23840
+    _subr__u64toa        = __native_entry__() + 4176
+    _subr__unquote       = __native_entry__() + 7232
+    _subr__validate_one  = __native_entry__() + 23664
+    _subr__value         = __native_entry__() + 13680
+    _subr__vnumber       = __native_entry__() + 17888
+    _subr__vsigned       = __native_entry__() + 19440
+    _subr__vstring       = __native_entry__() + 15760
+    _subr__vunsigned     = __native_entry__() + 19792
 )
 
 const (
     _stack__f32toa = 64
     _stack__f64toa = 80
+    _stack__get_by_path = 232
     _stack__html_escape = 64
     _stack__i64toa = 16
     _stack__lspace = 8
@@ -40,6 +43,7 @@ const (
     _stack__skip_number = 72
     _stack__skip_object = 128
     _stack__skip_one = 128
+    _stack__skip_one_fast = 144
     _stack__u64toa = 8
     _stack__unquote = 72
     _stack__validate_one = 128
@@ -53,6 +57,7 @@ const (
 var (
     _ = _subr__f32toa
     _ = _subr__f64toa
+    _ = _subr__get_by_path
     _ = _subr__html_escape
     _ = _subr__i64toa
     _ = _subr__lspace
@@ -61,6 +66,7 @@ var (
     _ = _subr__skip_number
     _ = _subr__skip_object
     _ = _subr__skip_one
+    _ = _subr__skip_one_fast
     _ = _subr__u64toa
     _ = _subr__unquote
     _ = _subr__validate_one
@@ -74,6 +80,7 @@ var (
 const (
     _ = _stack__f32toa
     _ = _stack__f64toa
+    _ = _stack__get_by_path
     _ = _stack__html_escape
     _ = _stack__i64toa
     _ = _stack__lspace
@@ -82,6 +89,7 @@ const (
     _ = _stack__skip_number
     _ = _stack__skip_object
     _ = _stack__skip_one
+    _ = _stack__skip_one_fast
     _ = _stack__u64toa
     _ = _stack__unquote
     _ = _stack__validate_one