… .github/workflows/benchmark-linux-x64.yml
… .github/workflows/codeql-analysis.yml
… .github/workflows/license-check.yml
… .github/workflows/push-check-go118.yml
… .github/workflows/push-check-linux-arm64.yml
… .github/workflows/push-check-linux-x64.yml
… .github/workflows/push-check-qemu.yml
… .github/workflows/push-check-windows.yml
… ast/search_test.go
… ast/testdata_test.go
… external_jsonlib_test/benchmark_test/search_test.go
… external_jsonlib_test/go.sum
… fuzz/ast_fuzz_test.go
… fuzz/fuzz_test.go
… internal/native/avx/native_amd64.s
… internal/native/avx/native_amd64_test.go
… internal/native/avx2/native_amd64.s
… internal/native/avx2/native_amd64_test.go
… internal/native/native_amd64_test.tmpl
… internal/native/sse/native_amd64.s
… internal/native/sse/native_amd64_test.go
… native/parsing.c
… native/scanning.c
… native/test/xprintf.h
… search_test.go