# Main Changes
use SIMD to accelerate the skip_object/array/number APIs in native code
Limits: not validate any more for performance, e.g. not validate the number format.

# Compare with master branch: 

## Common case, Performance x9

```
go: added golang.org/x/perf v0.0.0-20220920022801-e8d778a60d07
name             old time/op    new time/op    delta
GetOne_Sonic-12    12.8µs ± 4%     1.3µs ± 4%   -89.76%  (p=0.000 n=10+10)

name             old speed      new speed      delta
GetOne_Sonic-12  1.02GB/s ± 4%  9.93GB/s ± 4%  +876.39%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
GetOne_Sonic-12     24.0B ± 0%     24.0B ± 0%      ~     (all equal)

name             old allocs/op  new allocs/op  delta
GetOne_Sonic-12      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
```

## Bench with lots of compare operators, Performance x2
```
go: added golang.org/x/perf v0.0.0-20220920022801-e8d778a60d07
name                         old time/op    new time/op    delta
GetWithManyCompare_Sonic-12    7.82µs ± 8%    3.70µs ±13%   -52.66%  (p=0.000 n=9+10)

name                         old speed      new speed      delta
GetWithManyCompare_Sonic-12   126MB/s ±15%   271MB/s ±14%  +115.32%  (p=0.000 n=10+10)

name                         old alloc/op   new alloc/op   delta
GetWithManyCompare_Sonic-12     24.0B ± 0%     24.0B ± 0%      ~     (all equal)

name                         old allocs/op  new allocs/op  delta
GetWithManyCompare_Sonic-12      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
```

# Compare with others
```
goos: darwin
goarch: amd64
pkg: github.com/bytedance/sonic/external_jsonlib_test/benchmark_test
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkGetByKeys_Sonic-12               994468              1159 ns/op        11248.84 MB/s         24 B/op          1 allocs/op
BenchmarkGetByKeys_JsonParser-12           64651             19126 ns/op         681.53 MB/s           0 B/op          0 allocs/op
```

# Fuzz tests
```
fuzz: elapsed: 19m5s, execs: 2931657 (2495/sec), new interesting: 47 (total: 3278)
fuzz: elapsed: 19m8s, execs: 2942448 (3644/sec), new interesting: 47 (total: 3278)
fuzz: elapsed: 19m11s, execs: 2950775 (2797/sec), new interesting: 48 (total: 3279)
fuzz: elapsed: 19m14s, execs: 2959739 (2988/sec), new interesting: 48 (total: 3279)
fuzz: elapsed: 19m17s, execs: 2969151 (2923/sec), new interesting: 48 (total: 3279)
fuzz: elapsed: 19m20s, execs: 2976046 (2480/sec), new interesting: 48 (total: 3279)
fuzz: elapsed: 19m23s, execs: 2987277 (3743/sec), new interesting: 48 (total: 3279)
fuzz: elapsed: 19m26s, execs: 2996019 (2725/sec), new interesting: 48 (total: 3279)
fuzz: elapsed: 19m29s, execs: 3002092 (2176/sec), new interesting: 48 (total: 3279)
fuzz: elapsed: 19m32s, execs: 3009235 (2317/sec), new interesting: 48 (total: 3279)
fuzz: elapsed: 19m35s, execs: 3031807 (7516/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 19m38s, execs: 3040127 (2797/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 19m41s, execs: 3048943 (2998/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 19m44s, execs: 3058271 (3018/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 19m47s, execs: 3066288 (2710/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 19m50s, execs: 3072181 (1997/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 19m53s, execs: 3079643 (2455/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 19m56s, execs: 3089230 (3240/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 19m59s, execs: 3097342 (2704/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 20m2s, execs: 3105458 (2704/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 20m5s, execs: 3108273 (934/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 20m8s, execs: 3113499 (1751/sec), new interesting: 49 (total: 3280)
fuzz: elapsed: 20m11s, execs: 3118179 (1560/sec), new interesting: 49 (total: 3280)
^Cfuzz: elapsed: 20m14s, execs: 3122095 (1389/sec), new interesting: 50 (total: 3281)
--- PASS: FuzzMain (1214.70s)
```